Move PendingUpdate to commons so that is can be used from DatabaseHelper

+++ whois-update/src/test/java/net/ripe/db/whois/update/dao/JdbcPendingUpdateDaoTest.java
@@ -6,7 +6,7 @@ import net.ripe.db.whois.common.DateTimeProvider;
 import net.ripe.db.whois.common.dao.jdbc.domain.ObjectTypeIds;
 import net.ripe.db.whois.common.rpsl.ObjectType;
 import net.ripe.db.whois.common.rpsl.RpslObject;
-import net.ripe.db.whois.update.domain.PendingUpdate;
+import net.ripe.db.whois.common.domain.PendingUpdate;
 import org.hamcrest.core.Is;
 import org.joda.time.LocalDateTime;
 import org.junit.Before;
@@ -92,7 +92,7 @@ public class JdbcPendingUpdateDaoTest extends AbstractDaoTest {
         final PendingUpdate pending = new PendingUpdate(Sets.newHashSet("RouteIpAddressAuthentication"), object, dateTimeProvider.getCurrentDateTime());
         int pendingId = addPendingAuthentication(pending);
 
-        subject.remove(new PendingUpdate((long) pendingId, Sets.newHashSet("RouteIpAddressAuthentication"), object, new LocalDateTime()));
+        subject.remove(new PendingUpdate(pendingId, Sets.newHashSet("RouteIpAddressAuthentication"), object, new LocalDateTime()));
 
         final List<Map<String, Object>> result = databaseHelper.listPendingUpdates(object.getKey().toString());
         assertThat(result, hasSize(0));
